Add test infrastructure+integration test
Also reworked the repo structure a little bit
