+++ .github/workflows/testing_workflow.yml
@@ -1,0 +1,27 @@
+name: Run integration test
+
+on:
+  push:
+    branches:
+	- master
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.x'
+    - name: Install dependencies
+      run: |
+	python -m pip install --upgrade pip
+	pip install -r requirements.txt
+        pip install nose
+    - name: Run test
+      env:
+        CHROMEDRIVER_PATH=env.CHROMEWEBDRIVER
+      run: |
+        python nosetests
