+++ setup.cfg
@@ -1,0 +1,3 @@
+[nosetests]
+tests = test/
+
